producer/requirements.txt
flask==3.0.2
Werkzeug==3.0.2
kafka-python==2.0.2